forums de mots-clés ($afficher_texte = non) complètement plantés depuis le passage en 1.8 (Frédéric Taraud)
git-svn-id: svn://trac.rezo.net/spip/spip@4275 caf5f3e8-d4fe-0310-bb3e-c32d5e47d55d

+++ inc-messforum.php3
@@ -162,20 +162,22 @@ function enregistre_forum() {
 		exit;
 	}
 
-	// Verifier hash securite
-	include_ecrire("inc_admin.php3");
-	if (!verifier_action_auteur("ajout_forum $id_rubrique".
-	" $id_forum $id_article $id_breve".
-	" $id_syndic $alea", $hash)) {
-		spip_log('erreur hash forum');
-		die (_T('forum_titre_erreur')); 	# echec du POST
-	}
+	// Verifier hash securite pour les forums avec previsu
+	if ($GLOBALS['afficher_texte'] <> 'non') {
+		include_ecrire("inc_admin.php3");
+		if (!verifier_action_auteur("ajout_forum $id_rubrique".
+		" $id_forum $id_article $id_breve".
+		" $id_syndic $alea", $hash)) {
+			spip_log('erreur hash forum');
+			die (_T('forum_titre_erreur')); 	# echec du POST
+		}
 
-	// verifier fichier lock
-	$alea = preg_replace('/[^0-9]/', '', $alea);
-	if (!file_exists($hash = _DIR_SESSIONS."forum_$alea.lck"))
-		return $retour_forum; # echec silencieux du POST
-	unlink($hash);
+		// verifier fichier lock
+		$alea = preg_replace('/[^0-9]/', '', $alea);
+		if (!file_exists($hash = _DIR_SESSIONS."forum_$alea.lck"))
+			return $retour_forum; # echec silencieux du POST
+		unlink($hash);
+	}
 
 	// Entrer le message dans la base
 	$id_message = spip_abstract_insert('spip_forum', '(date_heure)', '(NOW())');
